+++ juriscraper/opinions/united_states/administrative_agency/asbca.py
@@ -21,20 +21,21 @@ class Site(OpinionSiteLinear):
         self.year = str(datetime.today().year)
         self.url = f"https://www.asbca.mil/Decisions/decisions{self.year}.html"
         self.status = "Published"
+        # Outdated headers may cause 403 responses. See #1693
         self.request["headers"] = {
             "accept": "text/html,application/xhtml+xml,application/xml;q=0.9,image/avif,image/webp,image/apng,*/*;q=0.8,application/signed-exchange;v=b3;q=0.7",
             "accept-encoding": "gzip, deflate, br, zstd",
             "accept-language": "en-US,en;q=0.9",
             "priority": "u=0, i",
-            "sec-ch-ua": '"Chromium";v="134", "Not:A-Brand";v="24", "Google Chrome";v="134"',
+            "sec-ch-ua": '"Chromium";v="142", "Not:A-Brand";v="24", "Google Chrome";v="142"',
             "sec-ch-ua-mobile": "?0",
             "sec-ch-ua-platform": '"Linux"',
             "sec-fetch-dest": "document",
             "sec-fetch-mode": "navigate",
-            "sec-fetch-site": "none",
             "sec-fetch-user": "?1",
+            "sec-fetch-site": "cross-site",
             "upgrade-insecure-requests": "1",
-            "user-agent": "Mozilla/5.0 (X11; Linux x86_64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/134.0.0.0 Safari/537.36",
+            "user-agent": "Mozilla/5.0 (X11; Linux x86_64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/142.0.0.0 Safari/537.36",
         }
         self.needs_special_headers = True
 
@@ -44,7 +45,7 @@ class Site(OpinionSiteLinear):
             self.year = "2024"
 
         rows = self.html.xpath(
-            "//tr[not(th) and not(.//span[@style='background-color:#F8C100;'])]"
+            "//tr[not(th) and not(.//span[@style='background-color:#F8C100;']) and descendant::a]"
         )
         for row in rows:
             if len(row.xpath(".//td")) != 4: