+++ juriscraper/scotus/scotus_docket_report_htm.py
@@ -156,6 +156,7 @@ class SCOTUSDocketReportHTM(SCOTUSDocketReportHTML):
 
         return {
             "date_filed": self.normalize_date(date_str),
+            "document_number": None,
             "description": description,
             "description_html": description_html,
             "attachments": [],